+++ src/main/java/wtf/norma/nekito/util/filter/filter/image/ShapeFilter.java
@@ -220,9 +220,9 @@ public class ShapeFilter extends WholeImageFilter {
 					int	g = (col & 0xFF00) >> 8;
 					int	b = (col & 0xFF);
 
-					r = (int)((sr*r/transp));
-					g = (int)((sg*g/transp));
-					b = (int)((sb*b/transp));
+					r = (sr*r/transp);
+					g = (sg*g/transp);
+					b = (sb*b/transp);
 				
 					// clip colors
 					if (r < 0)
